Changed class name and moved class file to controller package
