+++ mc-vip-lounge-client/pom.xml
@@ -10,9 +10,9 @@
   </parent>
 
   <groupId>com.mc.vip.lounge</groupId>
-  <artifactId>mc-vip-lounge-groupChat</artifactId>
+  <artifactId>mc-vip-lounge-messageRendering</artifactId>
   <version>1.0.0</version>
-  <name>mc-vip-lounge-groupChat</name>
+  <name>mc-vip-lounge-messageRendering</name>
 
   <url>http://maven.apache.org</url>
