+++ mc-vip-lounge-client/src/main/java/com/mc/vip/lounge/clientchat/chats/ChatRoom.java
@@ -1,0 +1,5 @@
+package com.mc.vip.lounge.clientchat.chats;
+
+public class ChatRoom {
+    
+}
