+++ mc-vip-lounge-client/src/main/java/com/mc/vip/lounge/main/ClientMain.java
@@ -2,7 +2,6 @@ package com.mc.vip.lounge.main;
 
 import javax.swing.*;
 
-import com.mc.vip.lounge.clientchat.chats.GroupChat;
 import com.mc.vip.lounge.clientchat.controller.ClientChatController;
 
 public class ClientMain {
